+++ Java Script Recall/README.md
@@ -1,22 +1,57 @@
 # React Repository Walkthrough
 
-Welcome to the walkthrough for exploring this folders in the repository. This guide will help you understand the structure and purpose of each folder and files.
+Welcome to the walkthrough for exploring this folders in the repository. This guide will help you understand the structure and purpose of important files and folder.
 
 ## Table of Contents
 
-- index.html
+- Java Script Recall\index.html
+- Java Script Recall\assets\scripts\
 
 ### Java Script Recall\index.html
-
 The HTML file provided serves as the foundation for a Java Script Recall project. Some important components and their purposes:
-- **Stylesheet**: The stylesheet contains styles and formatting rules for the HTML page.
-- **JavaScript File Deferment**: The commented section explains the use of the `defer` attribute in a `<script>` tag.
-- **JavaScript Module Import**: The commented section provides an example of importing a JavaScript module using the `<script>` tag with `type="module"`. It demonstrates how to import a JavaScript module from the specified file path.
+1. Stylesheet: The `<link>` tag includes a reference to an external CSS file. This stylesheet contains styles and formatting rules for the HTML page.
+
+2. JavaScript File Deferment: The commented section explains the use of the `defer` attribute in a `<script>` tag.
+
+3. JavaScript Module Import**: The commented section provides an example of importing a JavaScript module using the `<script>` tag with `type="module"`. It demonstrates how to import a JavaScript module from the specified file path.
+
+### Java Script Recall\assets\scripts\
+The "scripts" folder contains all the JavaScript files used in this project. It is recommended to follow the list below in order to understand the concepts and functionality from beginning to end.
+
+1. `app.js`: This is the main JavaScript file for the project and serves as the entry point. It demonstrates how to add a JavaScript file to an HTML page by linking it with a <script> tag. Additionally, it showcases the concept of importing and exporting modules in JavaScript.
+
+2. `ImportExport.js`: This file explores how to use the `import` and `export` keywords to share code between different JavaScript files. It is designed to be imported into the app.js file.
+
+3. `TypesOfVariable.js`: This file covers the different types of variables in JavaScript their differences and best practices for variable declaration.
+
+4. `ConstKeyword.js`: This file specifically focuses on the `const` keyword in JavaScript. It explains its usage, immutability, and how it differs from `let` and `var`.
+
+5. `JSOperators.js`: This file explores various JavaScript operators, including arithmetic, comparison, logical, assignment, and bitwise operators. It demonstrates how to use these operators effectively in different scenarios.
+
+6. `Functions.js`: This file covers the basics of JavaScript functions. It explains function declaration, invocation, parameters, return values.
+
+7. `ArrowFunction.js`: This file introduces arrow functions, a concise syntax for defining functions in JavaScript. It demonstrates how arrow functions can simplify function definitions.
+
+8. `ThisKeyword.js`: This file explains the `this` keyword in JavaScript. It covers its behavior within different contexts, such as object methods, event handlers, and constructors.
+
+9. `ObjectsAndClasses.js`: This file focuses on objects and classes in JavaScript. It explains how to create objects, define properties and methods, and utilize object-oriented principles using classes.
+
+10. `ArraysInJavaScript.js`: This file explores arrays in JavaScript, including array creation, manipulation, accessing elements, and common array methods.
+
+11. `DestructuringInJavaScript.js`: This file introduces destructuring assignment in JavaScript. It demonstrates how to extract values from objects and arrays using destructuring syntax.
+
+12. `SpreadOperatorInJavaScript.js`: This file explains the spread operator (`...`) in JavaScript. It showcases its usage for array manipulation, object merging, and function arguments.
+
+13. `ConditionalStatementsAndLoopsInJavaScript.js`: This file covers conditional statements (such as `if`, `else`, `switch`) and different types of loops (such as `for`, `while`, `do-while`) in JavaScript.
+
+14. `FunctionsAsValuesAndPassingFunctions.js`: This file explores the concept of functions as values in JavaScript. It explains how functions can be assigned to variables, passed as arguments, and returned from other functions.
+
+15. `FunctionsDefinedInsideOtherFunctions.js`: This file covers the concept of defining functions inside other functions, known as nested functions or closures.
+
+16. `PrimitivesAndReferenceValuesInJavaScript.js`: This file differentiates between primitive values (such as strings, numbers, booleans) and reference values (such as objects and arrays) in JavaScript. It covers the nuances of working with these value types.
 
 
 
-## Conclusion
+Feel free to explore and modify these files as needed, keeping in mind their interdependencies and the project's overall structure.
 
-In this walkthrough, we explored the various folders and files in the repository. Understanding the structure and purpose of these folders and files will help you navigate the codebase more efficiently.
-
-Feel free to explore the folder and files and make changes as needed. If you encounter any issues or have further questions, don't hesitate to reach out. Happy coding!
+Happy coding!
