+++ Java Script Recall/assets/scripts/ConstKeyword.js
@@ -1,0 +1,12 @@
+//---------------------------------------------------//
+console.log("Const keyword")
+
+//const is a keyword used to declare a variable that has a constant value, meaning its value cannot be reassigned or changed once it is assigned. It is short for "constant."
+//When you declare a variable using const, you must assign it a value at the time of declaration.
+const PI = 3.14159;
+console.log(PI); // 3.14159
+// Attempting to reassign the value will result in an error
+//PI = 3.14; // Error: Assignment to constant variable
+
+
+//---------------------------------------------------//
