+++ Java Script Recall/assets/scripts/JSOperators.js
@@ -1,0 +1,117 @@
+//---------------------------------------------------//
+console.log("JS Operators")
+
+//JavaScript supports various types of operators, including arithmetic, assignment, comparison, logical, and more. Here are some basic examples for each type:
+
+//----------------//
+//Arithmetic Operators:
+console.log("Arithmetic Operators")
+//----------------//
+
+//Addition: +
+let sum = 5 + 3; // 8
+console.log(sum)
+
+//Subtraction: -
+let difference = 10 - 3; // 7
+console.log(difference)
+
+//Multiplication: *
+let product = 4 * 5; // 20
+console.log(product)
+
+//Division: /
+let quotient = 15 / 3; // 5
+console.log(quotient)
+
+//Modulo (remainder): %
+let remainder = 17 % 4; // 1
+console.log(remainder)
+
+//----------------//
+//Assignment Operators:
+console.log("Assignment Operators")
+//----------------//
+
+//Assignment: =
+let x = 10;
+console.log(x)
+
+//Addition assignment: +=
+let y = 5;
+y += 3; // Equivalent to: y = y + 3; (y becomes 8)
+console.log(y)
+
+//Subtraction assignment: -=
+let z = 7;
+z -= 2; // Equivalent to: z = z - 2; (z becomes 5)
+console.log(z)
+
+//----------------//
+//Comparison Operators:
+console.log("Comparison Operators")
+//----------------//
+
+//Equal to: ==
+console.log(5 == 5); // true
+console.log(5 == '5'); // true (loose equality, converts types if needed)
+
+//Not equal to: !=
+console.log(5 != 3); // true
+console.log(5 != '5'); // false (loose inequality, converts types if needed)
+
+//Strict equal to: ===
+console.log(5 === 5); // true
+console.log(5 === '5'); // false (strict equality, no type conversion)
+
+//Strict not equal to: !==
+console.log(5 !== 3); // true
+console.log(5 !== '5'); // true (strict inequality, no type conversion)
+
+//----------------//
+//Logical Operators:
+console.log()
+//----------------//
+
+//Logical AND: &&
+console.log(true && false); // false
+
+//Logical OR: ||
+console.log(true || false); // true
+
+//Logical NOT: !
+console.log(!true); // false
+
+//----------------//
+//Unary Operators:
+console.log("Unary Operators")
+//----------------//
+
+//Increment: ++
+let count = 5;
+count++; // Equivalent to: count = count + 1; (count becomes 6)
+console.log(count)
+
+//Decrement: --
+let num = 8;
+num--; // Equivalent to: num = num - 1; (num becomes 7)
+console.log(num)
+
+//----------------//
+//String Operators:
+console.log("String Operators")
+//----------------//
+
+//String concatenation: +
+let greeting = 'Hello, ' + 'world!'; // "Hello, world!"
+console.log(greeting)
+
+//String concatenation assignment: +=
+let message = 'Welcome, ';
+message += 'John!'; // Equivalent to: message = message + 'John!'; (message becomes "Welcome, John!")
+console.log(message)
+
+//These are just a few examples of the different types of operators available in JavaScript. Operators allow you to perform various operations on values and variables, enabling you to manipulate and combine data within your code.
+
+
+//---------------------------------------------------//
