+++ Java Script Recall/index.html
@@ -12,6 +12,20 @@
         <script src="assets/scripts/app.js" defer></script>
     -->
     <script src = "assets/scripts/app.js" type="module"></script>
+    <script src = "assets/scripts/TypesOfVariable.js" type="module"></script>
+    <script src = "assets/scripts/ConstKeyword.js" type="module"></script>
+    <script src = "assets/scripts/JSOperators.js" type="module"></script>
+    <script src = "assets/scripts/Functions.js" type="module"></script>
+    <script src = "assets/scripts/ArrowFunction.js" type="module"></script>
+    <script src = "assets/scripts/ThisKeyword.js" type="module"></script>
+    <script src = "assets/scripts/ObjectsAndClasses.js" type="module"></script>
+    <script src = "assets/scripts/ArraysInJavaScript.js" type="module"></script>
+    <script src = "assets/scripts/DestructuringInJavaScript.js" type="module"></script>
+    <script src = "assets/scripts/SpreadOperatorInJavaScript.js" type="module"></script>
+    <script src = "assets/scripts/ConditionalStatementsAndLoopsInJavaScript.js" type="module"></script>
+    <script src = "assets/scripts/FunctionsAsValuesAndPassingFunctions.js" type="module"></script>
+    <script src = "assets/scripts/FunctionsDefinedInsideOtherFunctions.js" type="module"></script>
+    <script src = "assets/scripts/PrimitivesAndReferenceValuesInJavaScript.js" type="module"></script>
     <!--
         # JavaScript Module Import
         To import a JavaScript module using the script tag, you can use the `type="module"` attribute.
